CR comments from @jonas-ott

+++ tests/deprecated/test_modifier_regression_old.cpp
@@ -85,7 +85,7 @@ void test_verify_dist_poly5 (lfFixture *lfFix, gconstpointer data)
     float expected_y[] = {19.197506f, 497.0f, 933.42279f, 107.56808f};
 
     float coords [2];
-    for (long unsigned int i = 0; i < sizeof(x) / sizeof(float); i++)
+    for (size_t i = 0; i < sizeof(x) / sizeof(float); i++)
     {
         g_assert_true(mod->ApplyGeometryDistortion (x[i], y[i], 1, 1, coords));
         //g_print("\n%.8f, %.8f\n", coords[0], coords[1]);
@@ -116,7 +116,7 @@ void test_verify_dist_ptlens (lfFixture *lfFix, gconstpointer data)
     float expected_y[] = {19.339846f, 497.00046f, 927.90521f, 111.40639f};
 
     float coords [2];
-    for (long unsigned int i = 0; i < sizeof(x) / sizeof(float); i++)
+    for (size_t i = 0; i < sizeof(x) / sizeof(float); i++)
     {
         g_assert_true(mod->ApplyGeometryDistortion (x[i], y[i], 1, 1, coords));
         //g_print("\n%.8f, %.8f\n", coords[0], coords[1]);
@@ -146,7 +146,7 @@ void test_verify_vignetting_pa (lfFixture *lfFix, gconstpointer data)
     lf_u16 expected[] = {22406, 22406, 24156, 28803};
 
     lf_u16 coords [3] = {16000, 16000, 16000};
-    for (long unsigned int i = 0; i < sizeof(x) / sizeof(float); i++)
+    for (size_t i = 0; i < sizeof(x) / sizeof(float); i++)
     {
         g_assert_true(mod->ApplyColorModification(&coords[0], x[i], y[i], 1, 1, LF_CR_3(RED,GREEN,BLUE), 0));
         //g_print("\n%d, %d, %d\n", coords[0], coords[1], coords[2]);
@@ -182,7 +182,7 @@ void test_verify_subpix_linear (lfFixture *lfFix, gconstpointer data)
     };
 
     float coords [6];
-    for (long unsigned int i = 0; i < sizeof(x) / sizeof(float); i++)
+    for (size_t i = 0; i < sizeof(x) / sizeof(float); i++)
     {
         g_assert_true(mod->ApplySubpixelDistortion(x[i], y[i], 1, 1, coords));
         //g_print("{%.8f, %.8f, %.8f, %.8f, %.8f, %.8f},\n", coords[0], coords[1], coords[2], coords[3], coords[4], coords[5]);
@@ -216,7 +216,7 @@ void test_verify_subpix_poly3 (lfFixture *lfFix, gconstpointer data)
         {1270.11572266f, 99.91123199f, 1270.00000000f, 100.00000763f, 1269.96374512f, 100.02780914f}
     };
 
-    for (long unsigned int i = 0; i < sizeof(x) / sizeof(float); i++)
+    for (size_t i = 0; i < sizeof(x) / sizeof(float); i++)
     {
         float coords [6];
         g_assert_true(mod->ApplySubpixelDistortion (x[i], y[i], 1, 1, coords));
@@ -249,7 +249,7 @@ void test_verify_geom_fisheye_rectlinear (lfFixture *lfFix, gconstpointer data)
     float expected_y[] = {165.80287f, 497.00003f, 880.88129f, 191.18344f};
 
     float coords [2];
-    for (long unsigned int i = 0; i < sizeof(x) / sizeof(float); i++)
+    for (size_t i = 0; i < sizeof(x) / sizeof(float); i++)
     {
         g_assert_true(mod->ApplyGeometryDistortion (x[i], y[i], 1, 1, coords));
         //g_print("\n%.8f, %.8f\n", coords[0], coords[1]);
